Añadido ver comentarios en un post, cambios en los controladores , se añade hoja de estilos y se añade plantillas para añadir Juego
Se han modificado controladores
 y la plantilla para añadir un juego en caso de no encontrarlo, y se han reorganizado los controladores para mayor orden y limpieza.
Ligeras modificaciones en la plantilla y controladores de publicación que permiten que se vea los comentarios de una publicación.
Además se ha agregado una hoja de estilos general para la parte visual de la aplicación

+++ CheatHub/src/main/java/com/cheatHub/repositories/RepositorioComentario.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.cheatHub.entities.Comentario;
 import com.cheatHub.entities.Publicacion;
 @Repository
-public interface RepositorioComentario extends JpaRepository<Comentario,Long>{
+public interface RepositorioComentario extends JpaRepository<Comentario,Integer>{
 	List<Comentario> findByPublicacion(Publicacion publicacion);
 
 }
